updated medical officer
