+++ src/main/java/com/prison/Main.java
@@ -11,21 +11,17 @@ import java.util.Objects;
 
 public class Main extends Application {
 
+
     @Override
-    public void start(Stage stage) {
-        try {
-
-            Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getClassLoader().getResource("MedicalOfficerDashboard.fxml")));
-            Scene scene = new Scene(root);
-            stage.setScene(scene);
-            stage.show();
-
-        } catch(Exception e) {
-            e.printStackTrace();
-        }
+    public void start(Stage stage) throws IOException {
+        FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/com/prison/medical/MedicalOfficerDashboard.fxml"));
+        Scene scene = new Scene(fxmlLoader.load());
+        stage.setTitle("Hello!");
+        stage.setScene(scene);
+        stage.show();
     }
 
     public static void main(String[] args) {
-        launch(args);
+        launch();
     }
 }
