+++ src/main/java/com/prison/medical/InmateMedicalRecord.java
@@ -1,7 +1,38 @@
 package com.prison.medical;
 
+import javafx.event.ActionEvent;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+import java.util.Objects;
+
 public class InmateMedicalRecord {
 
+    private Stage stage;
+    private Scene scene;
+    private Parent root;
 
+    public void SwitchToInmateMedicalRecord(ActionEvent event) throws IOException {
 
+        root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("InmateMedicalRecord.fxml")));
+        stage = (Stage)((Node)event.getSource()).getScene().getWindow();
+        scene = new Scene(root);
+        stage.setScene(scene);
+        stage.show();
+
+    }
+
+    public void SwitchToMedicalOfficerDashboard(ActionEvent event) throws IOException {
+
+        root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("MedicalOfficerDashboard.fxml")));
+        stage = (Stage)((Node)event.getSource()).getScene().getWindow();
+        scene = new Scene(root);
+        stage.setScene(scene);
+        stage.show();
+
+    }
 }
