+++ src/main/java/com/prison/medical/MedicalInventoryRecord.java
@@ -1,38 +1,5 @@
 package com.prison.medical;
 
-import javafx.event.ActionEvent;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
+public class MedicalInventoryRecord{
 
-import java.io.IOException;
-import java.util.Objects;
-
-public class MedicalInventoryRecord {
-
-    private Stage stage;
-    private Scene scene;
-    private Parent root;
-
-    public void SwitchToInmateMedicalRecord(ActionEvent event) throws IOException {
-
-        root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("InmateMedicalRecord")));
-        stage = (Stage)((Node)event.getSource()).getScene().getWindow();
-        scene = new Scene(root);
-        stage.setScene(scene);
-        stage.show();
-
-    }
-
-    public void SwitchToMedicalOfficerDashboard(ActionEvent event) throws IOException {
-
-        root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("MedicalOfficerDashboard")));
-        stage = (Stage)((Node)event.getSource()).getScene().getWindow();
-        scene = new Scene(root);
-        stage.setScene(scene);
-        stage.show();
-
-    }
 }
