+++ src/main/resources/com/prison/medical/InmateMedicalRecord.fxml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.shape.*?>
-<?import javafx.scene.text.*?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.shape.Line?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="532.0" prefWidth="769.0" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.prison.medical.MedicalInventoryRecord">
+<AnchorPane prefHeight="532.0" prefWidth="769.0" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.prison.medical.InmateMedicalRecord">
    <children>
       <Text layoutX="20.0" layoutY="40.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Prison Management" wrappingWidth="199.21630859375">
          <font>
@@ -14,8 +15,8 @@
       </Text>
       <Line endX="755.0000610351562" layoutX="14.0" layoutY="62.0" startX="-14.0" />
       <Line endX="-44.5" endY="426.79290771484375" layoutX="217.0" layoutY="106.0" startX="-44.5" startY="-43.5" />
-      <Button layoutY="62.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="172.0" text="Dashboard" />
-      <Button layoutY="88.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="172.0" text="Inmate" />
+      <Button layoutY="62.0" mnemonicParsing="false" onAction="#SwitchToMedicalOfficerDashboard" prefHeight="26.0" prefWidth="172.0" text="Dashboard" />
+      <Button layoutY="88.0" mnemonicParsing="false" onAction="#SwitchToInmateMedicalRecord" prefHeight="26.0" prefWidth="172.0" text="Inmate" />
       <Button layoutY="114.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="172.0" text="Inventory" />
       <Button layoutY="140.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="172.0" text="Request" />
       <Button layoutY="166.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="172.0" text="Notification" />
