+++ src/main/resources/com/prison/medical/MedicalOfficerDashboard.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="534.0" prefWidth="813.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.prison.medical.MedicalOfficerDashboard">
+<AnchorPane prefHeight="534.0" prefWidth="813.0" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.prison.medical.MedicalOfficerDashboard">
    <children>
       <Text layoutX="14.0" layoutY="31.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Prison Management" wrappingWidth="198.02099609375">
          <font>
@@ -26,7 +26,7 @@
       <Text layoutX="525.0" layoutY="39.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Date:" />
       <DatePicker layoutX="562.0" layoutY="22.0" prefHeight="21.0" prefWidth="168.0" />
       <Accordion layoutX="227.0" layoutY="138.0" />
-      <Button layoutY="56.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="130.0" text="Dashboard" />
+      <Button layoutY="56.0" mnemonicParsing="false" onAction="#SwitchToMedicalOfficerDashboard" prefHeight="26.0" prefWidth="130.0" text="Dashboard" />
       <Button layoutY="134.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="130.0" text="Request" />
       <Button layoutY="108.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="130.0" text="Inventory" />
       <Button layoutY="82.0" mnemonicParsing="false" onAction="#SwitchToInmateMedicalRecord" prefHeight="26.0" prefWidth="130.0" text="Inmate" />
